+++ renderer/components/getters/Communities.tsx
@@ -1,0 +1,98 @@
+import React from 'react';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import { withStyles, Theme, createStyles, makeStyles } from '@material-ui/core/styles';
+import TableCell from '@material-ui/core/TableCell';
+import TableContainer from '@material-ui/core/TableContainer';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+import Paper from '@material-ui/core/Paper';
+import Box from '@material-ui/core/Box';
+import Container from '@material-ui/core/Container';
+import Typography from '@material-ui/core/Typography';
+import { Community, Contact } from '@d4data/archive-lib/src/types/schemas'
+import { GetterData } from '@d4data/archive-lib/src/types/standardizer/GetterReturn';
+import moment from 'moment';
+import ContactComponent from './Contacts/ContactComponent'
+
+const StyledTableCell = withStyles((theme: Theme) => createStyles({
+  head: {
+    backgroundColor: theme.palette.common.black,
+    color: theme.palette.common.white,
+  },
+  body: {
+    fontSize: 14,
+  },
+}))(TableCell);
+
+const StyledTableRow = withStyles((theme: Theme) => createStyles({
+  root: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.action.hover,
+    },
+  },
+}))(TableRow);
+
+const useStyles = makeStyles({
+  table: {
+    minWidth: 700,
+  },
+});
+
+export default function Communities({ data }: { data: NonNullable<GetterData<Array<Community>>> }) {
+  const classes = useStyles()
+  const [open, setOpen] = React.useState(false)
+  const [clickedProfile, setProfile] = React.useState<Contact | undefined>()
+
+  const handleClick = (profile: Contact) => {
+    setProfile(profile)
+    setOpen(true)
+  }
+  return (
+    <Container maxWidth="lg">
+      <Box my={ 4 }>
+        <Typography variant="h5" gutterBottom>
+          {`${ data.data.length } communities found`}
+        </Typography>
+      </Box>
+      <ContactComponent show={ open } onClose={ () => setOpen(false) } profile={ clickedProfile }/>
+      <Box my={ 2 }>
+        <TableContainer component={ Paper }>
+          <Table className={ classes.table } size="small" aria-label="a dense table">
+            <TableHead>
+              <StyledTableRow>
+                <StyledTableCell>Name</StyledTableCell>
+                <StyledTableCell>Joined date</StyledTableCell>
+              </StyledTableRow>
+            </TableHead>
+            <TableBody>
+              {data.data.map((row) => {
+                const profile = row;
+                return (
+                  <StyledTableRow key={ row.nickname }>
+                    <TableCell
+                      onClick={ () => handleClick(profile) }
+                      component="th"
+                      scope="row"
+                    >
+                      { profile.name }
+                    </TableCell>
+                    <TableCell
+                      onClick={ () => handleClick(profile) }
+                      component="th"
+                      scope="row"
+                    >
+                      { profile.joinedDate
+                        ? moment(new Date(profile.joinedDate)).format('yyyy-MM-DD')
+                        : 'No date provided' }
+                    </TableCell>
+                  </StyledTableRow>
+                )
+              })}
+            </TableBody>
+          </Table>
+        </TableContainer>
+      </Box>
+    </Container>
+  );
+}
